requirements.txt
torch==2.2.1
torchvision==0.17.1
efficientnet-pytorch==0.7.1
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.65.0
gradio==4.19.2 